:sparkles: feat(backend): feed resource에 있는 medias 관련 medias controller로 이동 (#273)
