+++ backend/src/api/medias/medias.controller.ts
@@ -18,6 +18,7 @@ import { LoggingInterceptor } from '@/common/interceptors/logging.interceptor';
 import { TimeoutInterceptor } from '@/common/interceptors/timeout.interceptor';
 import { ERROR_FILE_NOT_FOUND } from '@/constants/business-error';
 import {
+	CreateBodyImageMulterOptions,
 	CreateMemberCoverImageMulterOptions,
 	CreateMemberProfileImageMulterOptions,
 } from '@/utils/upload-media';
@@ -64,6 +65,26 @@ export class MediasController {
 		FilesInterceptor('files', 1, CreateMemberCoverImageMulterOptions()),
 	)
 	async postUploadCoverImage(@UploadedFiles() files: Express.MulterS3.File[]) {
+		if (!files?.length) {
+			throw BadRequestServiceException(ERROR_FILE_NOT_FOUND);
+		}
+		const locations = files.map(({ location }) => location);
+		return locations;
+	}
+
+	/**
+	 * @summary 피드 미디어 업로드
+	 *
+	 * @tag medias
+	 * @param files 업로드 파일 배열
+	 * @author YangGwangSeong <soaw83@gmail.com>
+	 * @returns 업로드 된 파일 배열
+	 */
+	@Post('/feeds')
+	@UseInterceptors(
+		FilesInterceptor('files', 10, CreateBodyImageMulterOptions()),
+	)
+	async postUploadFeedMedias(@UploadedFiles() files: Express.MulterS3.File[]) {
 		if (!files?.length) {
 			throw BadRequestServiceException(ERROR_FILE_NOT_FOUND);
 		}
